pyproject.toml
[project]
name = "rohans-weblog"
version = "0.1.0"
description = "Rohan Singh's Weblog - Django blog"
requires-python = ">=3.10"
dependencies = [
    "black>=25.12.0",
    "bleach[css]>=6.0",
    "dj-database-url>=2.1.0",
    "django>=5.0",
    "gunicorn>=21.0",
    "markdown>=3.5",
    "psycopg2-binary>=2.9.0",
    "pymdown-extensions>=10.0",
    "python-frontmatter>=1.0",
    "pygments>=2.0",
]
